3799: Extract EntityNodesTest
